fixup(journal): Unit 5 round-1 nits + NoTimeFilter sentinel warning
- renderJournalSinceContent: drop defensive `i < len(bodies)` guard;
  bodies is contract-guaranteed parallel to entries by read.Since.
- Lift skipIfNearUTCMidnight to untagged journal_test_helpers_test.go
  and adopt from journal_since_test.go's 5 call sites. Slow-tagged
  callers in journal_read_e2e_list_test.go / journal_read_e2e_since_test.go
  resolve the same symbol (untagged is visible to all builds).
- list.go: document NoTimeFilter's sentinel window contract
  (meta.EffectiveSince=zero, EffectiveUntil=9999-12-31 are NOT
  real timestamps — callers must not surface them in envelopes).

SageOx-Session: https://sageox.ai/repo/repo_019c5812-01e9-7b7d-b5b1-321c471c9777/sessions/2026-04-13T17-33-galexy-OxKTzU/view
