+++ cmd/ox/journal_read_e2e_list_test.go
@@ -97,22 +97,9 @@ func dayCeilUTC(t time.Time) time.Time {
 	return dayFloorUTC(t).Add(24 * time.Hour)
 }
 
-// skipIfNearUTCMidnight skips the test if `now` is within 10 seconds of
-// a UTC day boundary. The harness captures `now` once and the production
-// subprocess captures its own `time.Now()` a few milliseconds later;
-// when those straddle a day boundary the day-rounded window in the
-// envelope disagrees with the test's expected bounds. Ten seconds is far
-// more margin than the subprocess startup budget so the skip is rare in
-// practice.
-func skipIfNearUTCMidnight(t *testing.T, now time.Time) {
-	t.Helper()
-	u := now.UTC()
-	sod := time.Date(u.Year(), u.Month(), u.Day(), 0, 0, 0, 0, time.UTC)
-	eod := sod.Add(24 * time.Hour)
-	if u.Sub(sod) < 10*time.Second || eod.Sub(u) < 10*time.Second {
-		t.Skip("within 10s of UTC midnight — window bounds may straddle a day boundary")
-	}
-}
+// skipIfNearUTCMidnight lives in journal_test_helpers_test.go (untagged)
+// so both this slow-tagged file and the fast in-proc journal tests share
+// one source of truth.
 
 // TestJournalRead_JR01_MinimalDaily_List — one in-window daily.
 // Failure prevented: reader cannot locate memory/daily/ under the
